vmware_guest folder parameter is None by default
This change also affects the way that paths are built when looking up
VMs. Previously, the handling of slashes was not consistent and thus
required special care in multiple functions. Instead, this change
always keeps the original user argument intact, and strips all leading
and trailing slashes from the folder before attempting to build a path
with it. This makes the path building logic much simpler.

There is still some question about whether dcpath is needed. Whenever
I attempted to use the dcpath, the resulting call to
searchIndex.FindByInventoryPath would fail.