vmware: Only add configured network interfaces to facts
This change effectively filters out any network interfaces which were
not explicitly configured for the guest. This fixes some unexpected
behavior where a machine with multiple IP addresses (for example, when
Docker is installed, an internal IPv4 interface is added to
communicate with the container) would show one of the internal
addresses in the 'ipv4' field, but then no other information about the
corresponding hardware interface.